Update to 4.0.3
Fixed $() so that if the user bases in an array of garbage and attempts
to do DOM operations on its contents, ChocolateChipJS will convert it
to an empty array.
Changed how $(selector).prop works. It now set the element property to
true or false and returns true or false when you get a property.
Added $(selector).removeProp.
Updated TypeScript declaration file to reflect new behavior of
[].prop().